consolidate auth file
